% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome-pdf.r
\name{chrome_dump_pdf}
\alias{chrome_dump_pdf}
\title{"Print" to PDF}
\usage{
chrome_dump_pdf(url, path = NULL, overwrite = TRUE,
  chrome_bin = Sys.getenv("HEADLESS_CHROME"))
}
\arguments{
\item{url}{URL to read from}

\item{path}{path (with optional output filename) for the generated PDF. If \code{NULL} then
and \code{overwrite} is \code{FALSE}, the fuction will will ensure a uniquely-named file is
placed in the current working directory by incrementing trailing numbers before
the end of it.}

\item{overwrite}{overwrite existing file? Default: \code{TRUE}}

\item{chrome_bin}{the path to Chrome (auto-set from \code{HEADLESS_CHROME} environment variable)}
}
\value{
output fileame (invisibly)
}
\description{
"Print" to PDF
}
\note{
The default Chrome filename is \code{output.pdf}
}
\examples{
chrome_dump_pdf("https://www.r-project.org/")
}
