% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome-shot.r
\name{chrome_shot}
\alias{chrome_shot}
\title{Capture a screenshot}
\usage{
chrome_shot(url, width = NULL, height = NULL, path = NULL,
  overwrite = TRUE, chrome_bin = Sys.getenv("HEADLESS_CHROME"))
}
\arguments{
\item{url}{URL to read from}

\item{width, height}{screen size to emulate}

\item{path}{path (with optional output filename) for the generated PDF. If \code{NULL} then
and \code{overwrite} is \code{FALSE}, the fuction will will ensure a uniquely-named file is
placed in the current working directory by incrementing trailing numbers before
the end of it.}

\item{overwrite}{overwrite existing file? Default: \code{TRUE}}

\item{chrome_bin}{the path to Chrome (auto-set from \code{HEADLESS_CHROME} environment variable)}
}
\value{
\code{magick}
}
\description{
For the moment, the capture file is in the current working directory and named
\code{screenshot.png}. This will change, soon.
}
\details{
A \code{magick} image object is returned.
}
\note{
The default Chrome filename is \code{screenshot.png}
}
\examples{
chrome_shot("https://www.r-project.org/logo/Rlogo.svg")
}
