% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.r
\name{chrome_version}
\alias{chrome_version}
\title{Get Chrome version}
\usage{
chrome_version(chrome_bin = Sys.getenv("HEADLESS_CHROME"))
}
\arguments{
\item{chrome_bin}{the path to Chrome (auto-set from \code{HEADLESS_CHROME} environment variable)}
}
\description{
Get Chrome version
}
\examples{
chrome_version()
}
