% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headless.r
\name{decapitate}
\alias{decapitate}
\title{Run headless}
\usage{
decapitate(port = 9222L, url = NULL)
}
\arguments{
\item{port}{remote debugging port (defaults to \code{9222})}

\item{url}{visit this URL on start}
}
\description{
Run headless
}
