% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decapitated-package.R
\docType{package}
\name{decapitated}
\alias{decapitated}
\alias{decapitated-package}
\title{Headless 'Chrome' Orchestration}
\description{
The 'Chrome' browser \url{https://www.google.com/chrome/} has a headless mode
which can be instrumented programmatically. Tools are provided to perform headless
Chrome' instrumentation on the command-line, including retrieving the javascript-executed
web page, PDF output or screen shot of a URL.
}
\section{Important}{


You'll need to set an envrionment variable \code{HEADLESS_CHROME} to one of these two values:
\itemize{
\item Windows(32bit): \code{C:\\Program Files\\Google\\Chrome\\Application\\chrome.exe}
\item Windows(64bit): \code{C:\\Program Files (x86)/Google\\Chrome\\Application\\chrome.exe}
\item macOS: \code{/Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome}
\item Linux: \code{/usr/bin/google-chrome}
}

A guess is made (but not verified yet) if \code{HEADLESS_CHROME} is non-existent.

Use \code{~/.Renviron} to store this value for the time being.
}

\author{
Bob Rudis (bob@rud.is)
}
