% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-chromium.R
\name{download_chromium}
\alias{download_chromium}
\title{Download a standalone version of Chromium (recommended)}
\usage{
download_chromium(path = NULL, chromium_revision = 579032)
}
\arguments{
\item{path}{where to store the Chromium app.}

\item{chromium_revision}{version to download (keep default unless you require
a specific version)}
}
\description{
Download a standalone version of Chromium (recommended)
}
