% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto.R
\name{gep_active}
\alias{gep_active}
\title{Get test whether the gepetto server is active}
\usage{
gep_active(gep)
}
\arguments{
\item{gep}{a gepetto connection object}
}
\value{
logical (\code{TRUE} if alive)
}
\description{
Get test whether the gepetto server is active
}
\examples{
\dontrun{
gepetto() \%>\%
  gep_active()
}
}
