% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto.R
\name{gep_debug}
\alias{gep_debug}
\title{Get "debug-level" information of a running gepetto server}
\usage{
gep_debug(gep)
}
\arguments{
\item{gep}{a gepetto connection object}
}
\value{
debug info
}
\description{
Get "debug-level" information of a running gepetto server
}
\examples{
\dontrun{
gepetto() \%>\%
  gep_debug() \%>\%
  str()
}
}
