% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto.R
\name{gep_render_har}
\alias{gep_render_har}
\title{Render a page in a javascript context and serialize to HAR}
\usage{
gep_render_har(gep, url, width = 1440, height = 5000)
}
\arguments{
\item{gep}{a gepetto connection object}

\item{url}{the URL to fetch and render}

\item{width, height}{viewport width/height}
}
\value{
HAR
}
\description{
TODO: Modify the \code{puppeteer-har} node module to allow for saving content
}
\note{
content is not returned, just HAR information
}
\examples{
\dontrun{
gepetto(port=3000) \%>\%
  gep_render_har("https://r-project.org/")
}
}
