% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto.R
\name{gep_render_html}
\alias{gep_render_html}
\title{Render a page in a javascript context and serialize to HTML}
\usage{
gep_render_html(gep, url, width = 1440, height = 5000)
}
\arguments{
\item{gep}{a gepetto connection object}

\item{url}{the URL to fetch and render}

\item{width, height}{viewport width/height}
}
\value{
HTML
}
\description{
Render a page in a javascript context and serialize to HTML
}
\examples{
\dontrun{
gepetto(port=3000) \%>\%
  gep_render_html("https://r-project.org/")
}
}
