% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto.R
\name{gep_render_magick}
\alias{gep_render_magick}
\title{Render a page in a javascript context and take a screenshot}
\usage{
gep_render_magick(gep, url, width = 1440, height = 5000)
}
\arguments{
\item{gep}{a gepetto connection object}

\item{url}{the URL to fetch and render}

\item{width, height}{viewport width/height}
}
\value{
\code{magick} image
}
\description{
Render a page in a javascript context and take a screenshot
}
\examples{
\dontrun{
gepetto(port=3000) \%>\%
  gep_render_magick("https://r-project.org/")
}
}
