% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto.R
\name{gep_render_pdf}
\alias{gep_render_pdf}
\title{Render a page in a javascript context and rendero to PDF}
\usage{
gep_render_pdf(gep, url, path = NULL, overwrite = TRUE, width = 1440,
  height = 5000)
}
\arguments{
\item{gep}{a gepetto connection object}

\item{url}{the URL to fetch and render}

\item{path}{directory & filename to save the PDF to. If \code{NULL} will be saved
to a tempfile and it location will be returned.}

\item{overwrite}{if \code{TRUE} any existing \code{path} (file) will be overwritten}

\item{width, height}{viewport width/height}
}
\value{
object
}
\description{
Render a page in a javascript context and rendero to PDF
}
\examples{
\dontrun{
gepetto(port=3000) \%>\%
  gep_render_pdf("https://r-project.org/")
}
}
