% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto.R
\name{gep_stop}
\alias{gep_stop}
\title{Gracefully stop a gepetto instance}
\usage{
gep_stop(gep)
}
\arguments{
\item{gep}{a gepetto connection object}
}
\description{
Gracefully stop a gepetto instance
}
\examples{
\dontrun{
gepetto() \%>\%
  gep_stop()
}
}
