% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto.R
\name{gepetto}
\alias{gepetto}
\title{Create a connection to a Gepetto API server}
\usage{
gepetto(host = "localhost", port = 8080)
}
\arguments{
\item{host}{where is it running? Defaults to "\code{localhost}"}

\item{port}{same, but what port? Defaults to \code{8080} since the most common
use case is that you have \code{gepetto} running in a Docker container.
Use \code{3000} if you're running it locally via \code{npm} as that's the default
port for development.`}
}
\value{
A \code{gepetto} connection object
}
\description{
Create a connection to a Gepetto API server
}
\examples{
\dontrun{
gepetto()
}
}
