% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{get_chrome_env}
\alias{get_chrome_env}
\title{get an envrionment variable \code{HEADLESS_CHROME}}
\usage{
get_chrome_env()
}
\description{
get an envrionment variable \code{HEADLESS_CHROME}
}
\note{
This only return an envrionment variable \code{HEADLESS_CHROME}.
}
\examples{
get_env()
}
