% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto-helpers.R
\name{install_gepetto}
\alias{install_gepetto}
\title{Install gepetto}
\usage{
install_gepetto()
}
\description{
This function verifies you have what you need to install \code{gepetto} then
does so (globally). Ideally, you'll have \code{Node.js} and \code{npm} already on your
system (since this function will not install them).
}
\details{
While you \emph{can} try to install gepetto this way, it is highly recommended that
you follow the instructions (\url{https://gitlab.com/hrbrmstr/gepetto}) manually
since many Node.js/npm installations have permissions issues due to errant
historical use of \code{sudo}.
}
