% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{set_env}
\alias{set_env}
\title{set an envrionment variable \code{HEADLESS_CHROME}}
\usage{
set_env(env = Sys.getenv("HEADLESS_CHROME"))
}
\arguments{
\item{env}{path of chrome execute file for an envrionment variable \code{HEADLESS_CHROME}}
}
\description{
set an envrionment variable \code{HEADLESS_CHROME}
}
\note{
This only grabs the \code{<body>} \code{innerHTML} contents
}
\examples{
set_env("C:/Program Files/Google/Chrome/Application/chrome.exe")
}
