% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepetto-helpers.R
\name{start_gepetto}
\alias{start_gepetto}
\alias{stop_gepetto}
\title{Start/stop gepetto}
\usage{
start_gepetto(host = "localhost", port = 8080)

stop_gepetto(pxobj)
}
\arguments{
\item{host}{IP/host to use instead of \code{localhost}}

\item{port}{port to use besides the default (\code{8080})}

\item{pxobj}{the \code{processx} object of a running gepetto instance generated by \code{\link[=start_gepetto]{start_gepetto()}}}
}
\value{
\code{\link[=start_gepetto]{start_gepetto()}} returns a \code{processx} object
}
\description{
These functions can help start/stop gepetto instances. You must remember to
save the object returned by \code{\link[=start_gepetto]{start_gepetto()}}.
}
\details{
Note that \code{\link[=stop_gepetto]{stop_gepetto()}} kills the process. You can/should consider using
\code{\link[=gep_stop]{gep_stop()}} instead.
}
\examples{
\dontrun{
pid <- start_gepetto()
# ... do stuff
stop_gepetto(pid)
}
}
