% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-html.r
\name{chrome_dump_pdf}
\alias{chrome_dump_pdf}
\title{"Print" to PDF}
\usage{
chrome_dump_pdf(url)
}
\arguments{
\item{url}{URL to read from}
}
\description{
"Print" to PDF
}
\note{
this is a quick version of the function and will overwrite \code{output.pdf} if it exists in CWD
}
\examples{
chrome_dump_pdf("https://www.r-project.org/")
}
