% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-html.r
\name{chrome_read_html}
\alias{chrome_read_html}
\title{Read a URL via headless Chrome and return the renderd \code{<body>} \code{innerHTML} DOM elements}
\usage{
chrome_read_html(url)
}
\arguments{
\item{url}{URL to read from}
}
\description{
Read a URL via headless Chrome and return the renderd \code{<body>} \code{innerHTML} DOM elements
}
\note{
This only grabs the \code{<body>} \code{innerHTML} contents
}
\examples{
chrome_read_html("https://www.r-project.org/")
}
