% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome-shot.r
\name{chrome_shot}
\alias{chrome_shot}
\title{Capture a screenshot}
\usage{
chrome_shot(url, width = NULL, height = NULL, path = NULL,
  work_dir = NULL, chrome_bin = Sys.getenv("HEADLESS_CHROME"))
}
\arguments{
\item{url}{URL to read from}

\item{width, height}{screen size to emulate}

\item{path}{path (with optional output filename) for the generated PDF. If \code{NULL} then
and \code{overwrite} is \code{FALSE}, the fuction will will ensure a uniquely-named file is
placed in the current working directory by incrementing trailing numbers before
the end of it.}

\item{work_dir}{See special Section.}

\item{chrome_bin}{the path to Chrome (auto-set from \code{HEADLESS_CHROME} environment variable)}

\item{overwrite}{overwrite existing file? Default: \code{TRUE}}

\item{prime}{if \code{TRUE} preliminary URL retrieval requests will be sent to "prime" the
headless Chrome cache. This seems to be necessary primarily on recent versions of macOS.
If numeric, that number of "prime" requests will be sent ahead of the capture request.
If \code{FALSE} no priming requests will be sent.}
}
\value{
\code{magick}
}
\description{
A \code{magick} image object is returned.
}
\note{
The default Chrome filename is \code{screenshot.png}
}
\section{Working around headless Chrome & OS security restrictions}{

Security restrictions on various operating systems and OS configurations can cause
headless Chrome execution to fail. As a result, headless Chrome operations should
use a special directory for \code{decapitated} package operations. You can pass this
in as \code{work_dir}. If \code{work_dir} is \code{NULL} a \code{.rdecapdata} directory will be
created in your home directory and used for the data, crash dumps and utility
directories for Chrome operations.\cr
\cr
\code{tempdir()} does not always meet these requirements (after testing on various
macOS 10.13 systems) as Chrome does some interesting attribute setting for
some of its file operations.
\cr
If you pass in a \code{work_dir}, it must be one that does not violate OS security
restrictions or headless Chrome will not function.
}

\examples{
chrome_shot("https://www.r-project.org/logo/Rlogo.svg")
}
