% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decapitated-package.R
\docType{package}
\name{decapitated}
\alias{decapitated}
\alias{decapitated-package}
\title{Headless 'Chrome' Orchestration}
\description{
The 'Chrome' browser \url{https://www.google.com/chrome/} has a headless mode
which can be instrumented programmatically. Tools are provided to perform headless
Chrome' instrumentation on the command-line, including retrieving the javascript-executed
web page, PDF output or screen shot of a URL.
}
\section{Working around headless Chrome & OS security restrictions}{

Security restrictions on various operating systems and OS configurations can cause
headless Chrome execution to fail. As a result, headless Chrome operations should
use a special directory for \code{decapitated} package operations. You can pass this
in as \code{work_dir}. If \code{work_dir} is \code{NULL} a \code{.rdecapdata} directory will be
created in your home directory and used for the data, crash dumps and utility
directories for Chrome operations.\cr
\cr
\code{tempdir()} does not always meet these requirements (after testing on various
macOS 10.13 systems) as Chrome does some interesting attribute setting for
some of its file operations.
\cr
If you pass in a \code{work_dir}, it must be one that does not violate OS security
restrictions or headless Chrome will not function.
}

\section{Important}{


You'll need to set an envrionment variable \code{HEADLESS_CHROME} to one of these two values:
\itemize{
\item Windows(32bit): \code{C:\\Program Files\\Google\\Chrome\\Application\\chrome.exe}
\item Windows(64bit): \code{C:\\Program Files (x86)/Google\\Chrome\\Application\\chrome.exe}
\item macOS: \code{/Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome}
\item Linux: \code{/usr/bin/google-chrome}
}

A guess is made (but not verified yet) if \code{HEADLESS_CHROME} is non-existent.

It's best to use \code{~/.Renviron} to store this value.
}

\author{
Bob Rudis (bob@rud.is)
}
